main

+++ BeOff/src/main/java/com/lec/spring/service/BoardService.java
@@ -248,9 +248,11 @@ public class BoardService {
 		if(pageRows == null) pageRows = C.PAGE_ROWS;   // session 에 없으면 기본값으로
 		session.setAttribute("page", page);  // 현재 페이지 (세션에 저장해두자)
 		
-		// page 의 데이터 읽어오기
-		//Page<Qna> pageWrites = writeRepository.findAll(PageRequest.of(page - 1, pageRows, Sort.by(Order.desc("id"))));
-		Page<Qna> pageWrites = writeRepository.findBySubjectContaining(search, PageRequest.of(page - 1, pageRows, Sort.by(Order.desc("id"))));
+		Page<Qna> pageWrites = 
+				writeRepository.findBySubjectContaining(
+						search, 
+						PageRequest.of(page - 1, pageRows, Sort.by(Order.desc("id")))
+						);
 		
 		long cnt = pageWrites.getTotalElements();   // 글 목록 전체의 개수
 		int totalPage = pageWrites.getTotalPages(); //총 몇 '페이지' 분량인가?
@@ -311,12 +313,14 @@ public class BoardService {
 			// 첨부파일 추가
 			addFiles(files, write.getId());
 			
-			// 삭제할 첨부파일들 삭제
-			for(Long fileId : delfile) {
-				Qnafile file = fileRepository.findById(fileId).orElse(null);
-				if(file != null) {
-					delFile(file);  // 물리적으로 삭제
-					fileRepository.delete(file);  // DB 에서 삭제
+			if (delfile != null) {
+				// 삭제할 첨부파일들 삭제
+				for(Long fileId : delfile) {
+					Qnafile file = fileRepository.findById(fileId).orElse(null);
+					if(file != null) {
+						delFile(file);  // 물리적으로 삭제
+						fileRepository.delete(file);  // DB 에서 삭제
+					}
 				}
 			}
